health-check endpoint refactor
